♻️refactor(analysis): 트랜잭션 self-invocation 해결 + 엔드포인트 복수형 수정
- AnalysisService → AnalysisTransactionService 분리 (Spring 프록시 우회 방지)
- /api/v1/analysis → /api/v1/analysis-sessions (REST 복수 명사 규칙)
